Move BearerTokenAuthenticationPolicy to Http namespace/policy.hpp
