+++ sdk/core/azure-core/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Breaking Changes
 - Removed `DateTime::operator Duration()`.
+- Moved `Azure::Core::BearerTokenAuthenticationPolicy`, defined in `azure/core/credentials.hpp` to `Azure::Core::Http` namespace in `azure/core/http/policy.hpp` header.
 
 
 ## 1.0.0-beta.3 (2020-11-11)
