+++ sdk/core/azure-core/src/http/bearer_token_authentication_policy.cpp
@@ -1,11 +1,11 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include <azure/core/credentials.hpp>
+#include <azure/core/http/policy.hpp>
 
-using namespace Azure::Core;
+using namespace Azure::Core::Http;
 
-std::unique_ptr<Http::RawResponse> BearerTokenAuthenticationPolicy::Send(
+std::unique_ptr<RawResponse> BearerTokenAuthenticationPolicy::Send(
     Context const& context,
     Http::Request& request,
     Http::NextHttpPolicy policy) const
