+++ sdk/storage/azure-storage-blobs/src/blob_container_client.cpp
@@ -84,7 +84,7 @@ namespace Azure { namespace Storage { namespace Blobs {
       policies.emplace_back(p->Clone());
     }
     policies.emplace_back(std::make_unique<Storage::Details::StoragePerRetryPolicy>());
-    policies.emplace_back(std::make_unique<Core::BearerTokenAuthenticationPolicy>(
+    policies.emplace_back(std::make_unique<Core::Http::BearerTokenAuthenticationPolicy>(
         credential, Storage::Details::StorageScope));
     policies.emplace_back(
         std::make_unique<Azure::Core::Http::TransportPolicy>(options.TransportPolicyOptions));
