+++ sdk/storage/azure-storage-files-datalake/src/datalake_file_system_client.cpp
@@ -121,7 +121,7 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
     }
 
     policies.emplace_back(std::make_unique<Storage::Details::StoragePerRetryPolicy>());
-    policies.emplace_back(std::make_unique<Core::BearerTokenAuthenticationPolicy>(
+    policies.emplace_back(std::make_unique<Core::Http::BearerTokenAuthenticationPolicy>(
         credential, Azure::Storage::Details::StorageScope));
     policies.emplace_back(
         std::make_unique<Azure::Core::Http::TransportPolicy>(options.TransportPolicyOptions));
